✨ feat: implement responsive design for mobile devices (#95)
* ✨ feat: Add MobileTabBar component and update SideBar and AppLayout

The code changes introduce a new file called "MobileTabBar" that displays a tab bar for mobile devices. It includes icons and titles for different tabs such as chat, market, and settings. The "TopActions" component in the "SideBar" module is updated to set the active tab to "chat" when clicked. The "AppLayout" component is modified to include the "MobileTabBar" component for mobile devices. The "common" translation file is updated to change the title of the "market" tab to "discover" and add a new title for the "settings" tab. The "SettingLayout" component is updated to use the "AppLayout" component.

Changes:
- Added MobileTabBar component
- Updated SideBar and AppLayout components
- Modified common translation file
- Updated SettingLayout component

* ✨ feat: Add mobile support and new components for chat feature

This commit introduces mobile support for various components in the chat feature of the web application. It also adjusts the layout for mobile devices and adds new components. Additionally, it includes styling adjustments and imports.

Changes:
- Added mobile support for chat components
- Adjusted layout for mobile devices
- Introduced new components for chat feature
- Made styling adjustments and imports

* ✨ feat: Implement responsive design for mobile devices

- Update "AppMobileLayout.tsx" to set overflow-x to hidden and overflow-y to auto. Set width to 100vw and height to 100vh for the container.
- Modify "Sidebar/index.tsx" to remove MobileNavBar import. Wrap the topic section in a Flexbox component and place mobile version code inside Drawer component.
- Remove Router import in "settings/mobile.tsx" and update navBar prop in AppMobileLayout component to only include MobileNavBar component.
- Update style for LogoThree component in "Banner/Hero.tsx" to remove marginBottom property.
- Modify "Banner/index.tsx" to include mobile prop in Banner component. Move Upload component above Link component.
- Import and use "welcome/mobile.tsx" in "welcome/index.page.tsx" for mobile version. Wrap Banner component in Mobile component if device is mobile.
- Add "welcome/mobile.tsx" file with WelcomeLayout component using AppMobileLayout component and MobileNavBar component as navBar prop.
- Update margin-bottom and line-height properties for title class in "welcome/style.ts".

* 🔀 chore: merge to master

* ✨ feat: 优化 Mobile 模式下 Topic 展示

---------

Co-authored-by: arvinxx <arvinx@foxmail.com>